Fix handling of first VAD timestamp (#77)
* Fix handling of first VAD timestamp

When the initial RTP timestamp is very far from 0, it is thought to be from the previous epoch of 32bit integers
after an integer wrap around. When we do not have any saved audio levels, we know we should start processing.

* VAD initializes with nil current_timestamp

This allows us to special-case nil as being in the :current epoch of
timestamp rollovers, to avoid the case where the initial packet is thought
to be the previous epoch.